+++ src/main/java/com/cmeza/spring/jdbc/repository/annotations/methods/JdbcProcedure.java
@@ -3,6 +3,8 @@ package com.cmeza.spring.jdbc.repository.annotations.methods;
 import com.cmeza.spring.jdbc.repository.annotations.JdbcLoggeable;
 import com.cmeza.spring.jdbc.repository.annotations.parameters.OutParameter;
 import com.cmeza.spring.jdbc.repository.mappers.JdbcRowMapper;
+import com.cmeza.spring.jdbc.repository.naming.NamingStrategy;
+import com.cmeza.spring.jdbc.repository.naming.NoOpNamingStrategy;
 import org.springframework.core.annotation.AliasFor;
 
 import java.lang.annotation.*;
@@ -30,4 +32,6 @@ public @interface JdbcProcedure {
 
     @AliasFor(annotation = JdbcLoggeable.class)
     boolean loggeable() default false;
+
+    Class<? extends NamingStrategy> parametersNamingStrategy() default NoOpNamingStrategy.class;
 }