+++ src/main/java/com/cmeza/spring/jdbc/repository/aware/JdbcRepositoryAware.java
@@ -1,5 +1,6 @@
 package com.cmeza.spring.jdbc.repository.aware;
 
+import com.cmeza.spring.jdbc.repository.naming.NamingStrategy;
 import com.cmeza.spring.jdbc.repository.repositories.template.JdbcRepositoryTemplate;
 import com.cmeza.spring.jdbc.repository.resolvers.JdbcPropertyResolver;
 import org.springframework.beans.factory.Aware;
@@ -14,5 +15,8 @@ public interface JdbcRepositoryAware extends Aware {
     }
 
     default void setJdbcRepositoryTemplate(Map<String, JdbcRepositoryTemplate> jdbcRepositoryTemplates){
+    }
+
+    default void setNamingStrategy(NamingStrategy namingStrategy) {
     }
 }