+++ src/main/java/com/cmeza/spring/jdbc/repository/naming/NoOpNamingStrategy.java
@@ -1,0 +1,8 @@
+package com.cmeza.spring.jdbc.repository.naming;
+
+public class NoOpNamingStrategy implements NamingStrategy {
+    @Override
+    public String parse(String origin) {
+        return origin;
+    }
+}