+++ CHANGELOG.md
@@ -22,10 +22,16 @@
 ### Added
 - Support for unnamed parameters
 
-## [1.0.2] - 2023-06-11
+## [1.0.2] - 2023-11-06
 ### Added
 - Support for projection without implementation
+
+## [1.0.3] - 2023-11-08
+### Added
+- Global Naming Strategy
+- Naming Strategy for JdbcFunction, JdbcInsert, JdbcProcedure and JdbcUpdate
 
 [1.0.0]: https://github.com/cmeza20/spring-boot-starter-jdbc-repository/tree/1.0.0
 [1.0.1]: https://github.com/cmeza20/spring-boot-starter-jdbc-repository/tree/1.0.1
 [1.0.2]: https://github.com/cmeza20/spring-boot-starter-jdbc-repository/tree/1.0.2
+[1.0.3]: https://github.com/cmeza20/spring-boot-starter-jdbc-repository/tree/1.0.3